+++ sentry/src/main/java/io/sentry/Attachment.java
@@ -12,7 +12,7 @@ public final class Attachment {
   private @Nullable byte[] bytes;
   private @Nullable String pathname;
   private final @NotNull String filename;
-  private @NotNull String contentType;
+  private final @NotNull String contentType;
 
   /**
    * We could use Files.probeContentType(path) to determine the content type of the filename. This
@@ -24,15 +24,29 @@ public final class Attachment {
   private static final String DEFAULT_CONTENT_TYPE = "application/octet-stream";
 
   /**
-   * Initializes an Attachment with bytes.
+   * Initializes an Attachment with bytes and a filename.
    *
    * @param bytes The bytes of file.
    * @param filename The name of the attachment to display in Sentry.
    */
   public Attachment(final @NotNull byte[] bytes, final @NotNull String filename) {
+    this(bytes, filename, DEFAULT_CONTENT_TYPE);
+  }
+
+  /**
+   * Initializes an Attachment with bytes, filename and content type.
+   *
+   * @param bytes The bytes of file.
+   * @param filename The name of the attachment to display in Sentry.
+   * @param contentType The content type of the attachment.
+   */
+  public Attachment(
+      final @NotNull byte[] bytes,
+      final @NotNull String filename,
+      final @NotNull String contentType) {
     this.bytes = bytes;
     this.filename = filename;
-    this.contentType = DEFAULT_CONTENT_TYPE;
+    this.contentType = contentType;
   }
 
   /**
@@ -59,9 +73,27 @@ public final class Attachment {
    * @param filename The name of the attachment to display in Sentry.
    */
   public Attachment(final @NotNull String pathname, final @NotNull String filename) {
+    this(pathname, filename, DEFAULT_CONTENT_TYPE);
+  }
+
+  /**
+   * Initializes an Attachment with a path, a filename and a content type.
+   *
+   * <p>The file located at the pathname is read lazily when the SDK captures an event or
+   * transaction not when the attachment is initialized. The pathname string is converted into an
+   * abstract pathname before reading the file.
+   *
+   * @param pathname The pathname string of the file to upload as an attachment.
+   * @param filename The name of the attachment to display in Sentry.
+   * @param contentType The content type of the attachment.
+   */
+  public Attachment(
+      final @NotNull String pathname,
+      final @NotNull String filename,
+      final @NotNull String contentType) {
     this.pathname = pathname;
     this.filename = filename;
-    this.contentType = DEFAULT_CONTENT_TYPE;
+    this.contentType = contentType;
   }
 
   /**
@@ -98,14 +130,5 @@ public final class Attachment {
    */
   public @NotNull String getContentType() {
     return contentType;
-  }
-
-  /**
-   * Sets the content type of the attachment. Default is "application/octet-stream".
-   *
-   * @param contentType the content type of the attachment.
-   */
-  public void setContentType(final @NotNull String contentType) {
-    this.contentType = contentType;
   }
 }